refactor(backend): split google sync lifecycle
