+++ packages/backend/src/auth/services/google/google.auth.service.ts
@@ -14,7 +14,7 @@ import { UserError } from "@backend/common/errors/user/user.errors";
 import { normalizeEmail } from "@backend/common/helpers/email.util";
 import mongoService from "@backend/common/services/mongo.service";
 import EmailService from "@backend/email/email.service";
-import syncService from "@backend/sync/services/sync.service";
+import googleSyncLifecycleService from "@backend/sync/services/lifecycle/google-sync-lifecycle.service";
 import { findCompassUserBy } from "@backend/user/queries/user.queries";
 import userService from "@backend/user/services/user.service";
 import userMetadataService from "@backend/user/services/user-metadata.service";
@@ -47,12 +47,14 @@ class GoogleAuthService {
   };
 
   private restartGoogleCalendarSyncInBackground = (cUserId: string) => {
-    syncService.restartGoogleCalendarSync(cUserId).catch((err) => {
-      logger.error(
-        `Something went wrong with starting calendar sync for user ${cUserId}`,
-        err,
-      );
-    });
+    googleSyncLifecycleService
+      .restartGoogleCalendarSync(cUserId)
+      .catch((err) => {
+        logger.error(
+          `Something went wrong with starting calendar sync for user ${cUserId}`,
+          err,
+        );
+      });
   };
 
   async googleSignup(
@@ -189,7 +191,7 @@ class GoogleAuthService {
     const googleOAuthClient = new GoogleOAuthClient();
     googleOAuthClient.oauthClient.setCredentials(oAuthTokens);
 
-    syncService
+    googleSyncLifecycleService
       .importIncremental(cUserId, googleOAuthClient.getGcalClient())
       .catch(async (err) => {
         if (
