+++ packages/backend/src/common/errors/handlers/error.express.handler.ts
@@ -24,7 +24,7 @@ import {
 } from "@backend/common/types/error.types";
 import { type SessionResponse } from "@backend/common/types/express.types";
 import { sseServer } from "@backend/servers/sse/sse.server";
-import syncService from "@backend/sync/services/sync.service";
+import googleSyncLifecycleService from "@backend/sync/services/lifecycle/google-sync-lifecycle.service";
 import { getSyncByToken } from "@backend/sync/util/sync.queries";
 import { findCompassUserBy } from "@backend/user/queries/user.queries";
 import userService from "@backend/user/services/user.service";
@@ -131,7 +131,7 @@ const handleGoogleError = async (
   }
 
   if (isFullSyncRequired(e)) {
-    syncService
+    googleSyncLifecycleService
       .restartGoogleCalendarSync(userId, { force: true })
       .catch((err) => {
         logger.error(
