+++ packages/web/src/views/Calendar/components/Draft/hooks/actions/submit.parser.test.ts
@@ -29,9 +29,11 @@ mock.module("@web/common/validators/grid.event.validator", () => ({
 
 mock.module("@web/common/validators/someday.event.validator", () => ({
   validateSomedayEvent,
+  validateSomedayEvents: mock((events: Schema_SomedayEvent[]) => events),
 }));
 
 mock.module("@web/common/utils/event/event.util", () => ({
+  assembleDefaultEvent: mock(async () => createMockGridEvent()),
   assembleGridEvent,
 }));
